# openapi.yaml
openapi: 3.0.3
info:
  title: AgentMD — Agentic Web-to-Markdown API
  version: 1.0.0
  description: |
    Convert a public web page to clean Markdown for AI agents.
    Pay-per-call with Solana USDC via HTTP 402 (x402 V2-shaped challenge,
    direct-to-wallet settlement, zero custody).

    This is a technical web-scraping / data-processing tool only.
    It is not a wallet, exchange, broker, or financial product.
  contact:
    name: AgentMD operator
servers:
  - url: https://your-app.onrender.com
    description: Production (set PUBLIC_BASE_URL)
  - url: http://localhost:8787
    description: Local development
paths:
...
  /v1/scrape:
    post:
      summary: Scrape a URL into Markdown (payment required)
      operationId: scrapeToMarkdown
      description: |
        Call without payment proof → HTTP 402 with `PAYMENT-REQUIRED` header
        (base64 PaymentRequired) and the same object as JSON body.

        Pay the listed USDC amount on Solana to `payTo`, then retry with either:
        - Header `PAYMENT-SIGNATURE`: base64 JSON `{ "x402Version": 2, "payload": { "transaction": "<txid>" } }`
        - Header `X-Payment-Tx`: raw Solana tx signature
        - Body field `paymentTxSignature`: raw Solana tx signature
      parameters:
        - name: PAYMENT-SIGNATURE
          in: header
          required: false
          schema: { type: string }
          description: Base64-encoded PaymentPayload-like JSON (AgentMD uses payload.transaction = txid).
        - name: X-Payment-Tx
          in: header
          required: false
          schema: { type: string }
          description: Raw Solana transaction signature.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrapeRequest"
            examples:
              unpaid:
                summary: First call (expect 402)
                value:
                  url: https://example.com
              paid:
                summary: Retry after on-chain payment
                value:
                  url: https://example.com
                  paymentTxSignature: 5VERv8NMvzbJMEkV8xnrLkEaWRtSz9CosKDYk6ZrL3jK...
      responses:
        "200":
          description: Clean Markdown payload
          headers:
            PAYMENT-RESPONSE:
              description: Base64 settlement summary
              schema: { type: string }
            Link:
              description: Discovery links to OpenAPI / x402 well-known
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeSuccess"
        "402":
          description: Payment required or verification failed
          headers:
            PAYMENT-REQUIRED:
              description: Base64 PaymentRequired (x402 V2-shaped)
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequired"
        "400":
          description: Missing/invalid URL or unsafe URL
        "413":
          description: Request body too large
        "415":
          description: Unsupported upstream content type
        "502":
          description: Upstream page fetch failed
        "504":
          description: Upstream fetch timed out
components:
  schemas:
    ScrapeRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          example: https://example.com
        paymentTxSignature:
          type: string
          description: Solana tx signature proving payment to payTo
    ScrapeSuccess:
      type: object
      required: [markdown, url, meta]
      properties:
        markdown: { type: string }
        url: { type: string, format: uri }
        title: { type: string, nullable: true }
        meta:
          type: object
          properties:
            contentType: { type: string, nullable: true }
            fetchedAt: { type: string, format: date-time }
            bytes: { type: integer }
            truncated: { type: boolean }
            payment:
              type: object
              properties:
                asset: { type: string }
                transaction: { type: string }
                explorerUrl: { type: string }
    PaymentRequired:
      type: object
      properties:
        x402Version: { type: integer, example: 2 }
        error: { type: string, example: PAYMENT_REQUIRED }
        resource:
          type: object
          properties:
            url: { type: string }
            description: { type: string }
            mimeType: { type: string }
        accepts:
          type: array
          items:
            $ref: "#/components/schemas/AcceptRequirement"
        agentmd:
          type: object
          description: Human-readable twin of payment facts (AgentMD extension)
    AcceptRequirement:
      type: object
      required: [scheme, network, amount, asset, payTo, maxTimeoutSeconds]
      properties:
        scheme: { type: string, example: exact }
        network:
          type: string
          description: CAIP-2 chain id
          example: solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp
        amount:
          type: string
          description: Atomic token units (USDC 6 decimals → 0.01 = 10000)
        asset:
          type: string
          description: SPL mint address (or "native" for SOL fallback)
        payTo: { type: string }
        maxTimeoutSeconds: { type: integer }
        extra: { type: object }